Crate communication_thread

Example on how to use a communication thread alongside with the GUI thread.

Tricks used here:

Use a channel to show data on the GUI.

Run an async function on the GUI event loop.

Use a separate thread to handle incoming data and put it into a channel.

Functions

build_ui

main

spawn_local_handler
Spawn channel receive task on the main event loop.

start_communication_thread
Spawn separate thread to handle communication.